requirements-API.txt
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
pymupdf>=1.23.0
Pillow>=10.0.0